+++ src/modules/discord-bot/discord-bot.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { DiscordBotController } from './controllers/discord-bot.controller';
+import { BotController } from 'src/common/controllers/bots.controller';
 import { MongooseModule } from '@nestjs/mongoose';
 import { Logs, logSchema } from './entities/discord-log-entity';
 import { DiscordService } from './services/discord-bot.service';
@@ -9,7 +9,7 @@ import { SentTipsService } from './services/send-tip.service';
   imports: [
     MongooseModule.forFeature([{ name: Logs.name, schema: logSchema }]),
   ],
-  controllers: [DiscordBotController],
+  controllers: [BotController],
   providers: [DiscordService, SentTipsService],
 })
 export class DiscordBotModule {}
