--- src/modules/discord-bot/dto/index.ts
@@ -1,2 +1,0 @@
-export * from '../dto/create-discord-tip.dto';
-export * from '../dto/update-discord-bot.dto';
