--- src/modules/discord-bot/dto/update-discord-bot.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateDiscordTipDto } from './create-discord-tip.dto';
-
-export class UpdateDiscordBotDto extends PartialType(CreateDiscordTipDto) {}
